requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.2.0.2
Flask-Script==2.0.5
Flask-WTF==0.10.2
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
gunicorn==19.1.1
itsdangerous==0.24
mixpanel-py3==3.0.4
